+++ crates/storage/nippy-jar/Cargo.toml
@@ -26,7 +26,7 @@ sucds = "~0.8"
 
 
 [dev-dependencies]
-rand = "0.8"
+rand = { version = "0.8", features = ["small_rng"] }
 
 
 [features]
